+++ tests/integration/Requests/PublicGroupRequestTest.php
@@ -4,6 +4,10 @@
 namespace Tests\Integration\Requests;
 
 
+use CTApi\Models\GroupCategory;
+use CTApi\Models\GroupInformation;
+use CTApi\Models\GroupPlace;
+use CTApi\Models\TargetGroup;
 use CTApi\Requests\PublicGroupRequest;
 use Tests\Integration\TestCaseAuthenticated;
 use Tests\Integration\TestData;
@@ -37,10 +41,20 @@ class PublicGroupRequestTest extends TestCaseAuthenticated
             }
         }
 
-        print_r($foundGroup);
-
         $this->assertNotNull($foundGroup);
         $this->assertEquals($this->groupName, $foundGroup->getName());
+
+        $this->assertNotNull($foundGroup->getInformation());
+        $this->assertInstanceOf(GroupInformation::class, $foundGroup->getInformation());
+
+        $this->assertNotNull($foundGroup->getInformation()->getGroupCategory());
+        $this->assertInstanceOf(GroupCategory::class, $foundGroup->getInformation()->getGroupCategory());
+
+        $this->assertNotNull($foundGroup->getInformation()->getTargetGroup());
+        $this->assertInstanceOf(TargetGroup::class, $foundGroup->getInformation()->getTargetGroup());
+
+        $this->assertNotNull($foundGroup->getInformation()->getGroupPlaces());
+        $this->assertIsArray($foundGroup->getInformation()->getGroupPlaces());
     }
 
 }